+++ micro_manager/micro_manager.py
@@ -483,16 +483,6 @@ class MicroManagerCoupling(MicroManager):
                 self._mesh_vertex_ids,
             ) = domain_decomposer.filter_duplicate_coords(all_coords, all_ids)
 
-            # Global coordinates that are necessary for model adaptivity
-            # TODO: Avoid the allgather by smartly selecting the relevant coordinates in model adaptivity
-            self._global_mesh_vertex_coords = self._comm.allgather(
-                self._mesh_vertex_coords
-            )
-
-        if not self._is_parallel or self._is_load_balancing:
-            # For a serial run or when load balancing, the local mesh vertex coordinates are the global mesh vertex coordinates
-            self._global_mesh_vertex_coords = self._mesh_vertex_coords
-
         if self._mesh_vertex_coords.size == 0:
             if self._is_parallel:
                 self._is_rank_empty = True